requirements.txt
# Основные
cassandra-driver>=3.25.0